More unit tests and minor improvements in oroject.

+++ Score-services/src/main/scala/com/scoring/dto/RuleDto.scala
@@ -1,3 +1,3 @@
 package com.scoring.dto
 
-case class RuleDto (ruleId: String, appliesOn: String, score:Int, enableFlag: Boolean, resourceName: String, ruleResource: List[Int])
+case class RuleDto (ruleId: String, appliesOn: String, score:Double, enableFlag: Boolean, resourceName: String, ruleResource: List[Int])
